pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nibot"
version = "1.0.0"
description = "Lightweight multi-channel AI agent framework"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.40",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "loguru>=0.7",
    "httpx>=0.27",
    "python-frontmatter>=1.0",
    "croniter>=1.3",
]

[project.optional-dependencies]
telegram = ["python-telegram-bot>=21.0"]
feishu = ["lark-oapi>=1.0"]
web = ["readability-lxml>=0.8", "lxml>=5.0"]
all = ["nibot[telegram,feishu,web]"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[project.scripts]
nibot = "nibot.__main__:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
